+++ src/modules/esm.mjs
@@ -1,0 +1,30 @@
+import path, { join } from 'path';
+import { release, version } from 'os';
+import { createServer as createServerHttp } from 'http';
+import './files/c.js';
+
+const random = Math.random();
+
+let unknownFile = random > 0.5 ? './files/a.json' : './files/b.json';
+
+const { default: unknownObject } = await import(unknownFile, { with: { type: 'json' } });
+
+console.log(`Release ${release()}`);
+console.log(`Version ${version()}`);
+console.log(`Path segment separator is "${path.sep}"`);
+
+console.log(`Path to current file is ${join(import.meta.filename)}`);
+console.log(`Path to current directory is ${join(import.meta.dirname)}`);
+
+const myServer = createServerHttp((_, res) => {
+	res.end('Request accepted');
+});
+
+const PORT = 3000;
+
+myServer.listen(PORT, () => {
+	console.log(`Server is listening on port ${PORT}`);
+	console.log('To terminate it, use Ctrl+C combination');
+});
+
+export { unknownObject, myServer };
